@@ -15,51 +15,50 @@
     "lint:ts": "eslint --fix",
     "lint:styles": "stylelint --fix src/**/*.scss",
     "nswag": "node_modules\\.bin\\nswag run nswag.json",
-    "codegen:run": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net9.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features",
-    "codegen:run-i18n": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net9.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features --enable-i18n true",
+    "codegen:run": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net9.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features --standalone-components true",
+    "codegen:run-i18n": "entry-codegen --source-assembly ../Enigmatry.Entry.Blueprint.CodeGeneration.Setup/bin/Debug/net9.0/Enigmatry.Entry.Blueprint.CodeGeneration.Setup.dll --destination-directory ./src/app/features --enable-i18n true --standalone-components true",
     "codegen:install": "dotnet tool install enigmatry.entry.codegeneration.tools -g",
     "codegen:update": "dotnet tool update enigmatry.entry.codegeneration.tools -g",
     "i18n": "ng extract-i18n --format json --output-path src/i18n --out-file messages.en.json & sortjson src/i18n"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.2.10",
-    "@angular/cdk": "^19.2.15",
-    "@angular/common": "^19.2.10",
-    "@angular/compiler": "^19.2.10",
-    "@angular/core": "^19.2.10",
-    "@angular/forms": "^19.2.10",
-    "@angular/material": "^19.2.15",
-    "@angular/material-date-fns-adapter": "^19.2.15",
-    "@angular/platform-browser": "^19.2.10",
-    "@angular/platform-browser-dynamic": "^19.2.10",
-    "@angular/router": "^19.2.10",
-    "@azure/msal-browser": "3.2.0",
-    "@enigmatry/entry-components": "19.0.0",
-    "@enigmatry/entry-form": "19.0.0",
-    "@microsoft/applicationinsights-web": "3.3.4",
-    "@ngx-formly/core": "6.3.12",
-    "@ngx-formly/material": "6.3.12",
-    "date-fns": "^3.6.0",
-    "rxjs": "^7.8.2",
+    "@angular/cdk": "^20.2.0",
+    "@angular/common": "^20.2.0",
+    "@angular/compiler": "^20.2.0",
+    "@angular/core": "^20.2.0",
+    "@angular/forms": "^20.2.0",
+    "@angular/material": "^20.2.0",
+    "@angular/material-date-fns-adapter": "^20.2.0",
+    "@angular/platform-browser": "^20.2.0",
+    "@angular/platform-browser-dynamic": "^20.2.0",
+    "@angular/router": "^20.2.0",
+    "@azure/msal-browser": "4.21.0",
+    "@enigmatry/entry-components": "20.0.1-preview.2",
+    "@enigmatry/entry-form": "20.0.1-preview.2",
+    "@microsoft/applicationinsights-web": "3.3.9",
+    "@ngx-formly/core": "7.0.0",
+    "@ngx-formly/material": "7.0.0",
+    "date-fns": "^4.1.0",
+    "rxjs": "^7.8.1",
     "tslib": "^2.8.1",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular/build": "^19.2.11",
-    "@angular/cli": "^19.2.11",
-    "@angular/compiler-cli": "^19.2.10",
-    "@angular/localize": "^19.2.10",
-    "@enigmatry/eslint-config": "19.0.0",
-    "@enigmatry/scss-foundation": "19.0.0",
-    "@enigmatry/stylelint-config": "19.0.0",
-    "@types/jest": "29.5.14",
-    "jest": "29.7.0",
+    "@angular/build": "^20.2.0",
+    "@angular/cli": "^20.2.0",
+    "@angular/compiler-cli": "^20.2.0",
+    "@angular/localize": "^20.2.0",
+    "@enigmatry/eslint-config": "20.0.1-preview.2",
+    "@enigmatry/scss-foundation": "20.0.1-preview.2",
+    "@enigmatry/stylelint-config": "20.0.1-preview.2",
+    "@types/jest": "30.0.0",
+    "jest": "30.0.5",
     "jest-junit": "^16.0.0",
-    "jest-preset-angular": "14.5.5",
-    "json-sort-cli": "^4.0.3",
-    "nswag": "^14.4.0",
+    "jest-preset-angular": "15.0.0",
+    "json-sort-cli": "^4.0.10",
+    "nswag": "^14.5.0",
     "run-script-os": "^1.1.6",
-    "typescript": "5.8.3"
+    "typescript": "5.9.2"
   }
 }
