@@ -1,10 +1,9 @@
-import { Component, ElementRef, OnInit } from '@angular/core';
+import { Component, ElementRef, inject, OnInit } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { SizeService } from '@services/size.service';
 import { MenuComponent } from '@shared/components/navigation/menu.component';
 
 @Component({
-  standalone: true,
   imports: [RouterOutlet, MenuComponent],
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -12,9 +11,8 @@ import { MenuComponent } from '@shared/components/navigation/menu.component';
 })
 export class AppComponent implements OnInit {
   title = 'enigmatry-entry-blueprint-app';
-
-  constructor(private readonly element: ElementRef,
-    private readonly sizeService: SizeService) { }
+  private readonly element: ElementRef = inject(ElementRef);
+  private readonly sizeService: SizeService = inject(SizeService);
 
   ngOnInit(): void {
     this.sizeService.startTrackingResizeOf(this.element);
