@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { AuthService } from '@app/auth/auth.service';
 import { ApplicationInsightsService } from '@services/application-insights.service';
 
@@ -8,9 +8,8 @@ export const initFactory = (service: AppInitService) => () => service.init();
   providedIn: 'root'
 })
 export class AppInitService {
-  constructor(
-    private authService: AuthService,
-    private appInsightsService: ApplicationInsightsService) { }
+  private readonly authService: AuthService = inject(AuthService);
+  private readonly appInsightsService: ApplicationInsightsService = inject(ApplicationInsightsService);
 
   init = async(): Promise<void> => {
     await this.authService.initialize();
