@@ -1,37 +1,34 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { GetProductsResponseItem, PermissionId, ProductsClient } from '@api';
+import { Component, inject, OnInit } from '@angular/core';
+import { MatButton } from '@angular/material/button';
+import { RouterModule } from '@angular/router';
+import { GetProductsResponseItem, PagedResponseOfGetProductsResponseItem, PermissionId, ProductsClient } from '@api';
 import { PermissionService } from '@app/auth/permissions.service';
+import { EntryPermissionModule, EntrySearchFilterModule } from '@enigmatry/entry-components';
 import { ContextMenuItem, PagedData } from '@enigmatry/entry-components/table';
 import { BaseListComponent } from '@shared/list-component/base-list-component.model';
 import { RouteSegments } from '@shared/model/route-segments';
-import { Observable, Subscription, switchMap, tap } from 'rxjs';
+import { lastValueFrom, Subscription, switchMap, tap } from 'rxjs';
+import { ProductsGeneratedModule } from '../generated/products-generated.module';
 import { GetProductsQuery } from '../models/get-products-query.model';
 
 @Component({
-  standalone: false,
+  imports: [EntrySearchFilterModule, RouterModule, ProductsGeneratedModule, EntryPermissionModule, MatButton],
   selector: 'app-product-list',
   templateUrl: './product-list.component.html',
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent extends BaseListComponent implements OnInit {
   data: PagedData<GetProductsResponseItem>;
-  query = new GetProductsQuery();
+  override query = new GetProductsQuery();
   PermissionId = PermissionId;
   contextMenuItems: ContextMenuItem[] = [];
-
-  constructor(private client: ProductsClient,
-    protected router: Router,
-    protected activatedRoute: ActivatedRoute,
-    private permissionService: PermissionService) {
-    super(router, activatedRoute);
-  }
+  private readonly permissionService: PermissionService = inject(PermissionService);
+  private readonly client: ProductsClient = inject(ProductsClient);
 
   ngOnInit(): void {
     this.watchQueryParamsAndGetProducts();
     this.initContextMenuItems();
   }
-
 
   readonly onRowSelected = async(rowData: GetProductsResponseItem) => {
     await this.router.navigate([RouteSegments.edit, rowData.id], { relativeTo: this.activatedRoute });
@@ -41,9 +38,8 @@ export class ProductListComponent extends BaseListComponent implements OnInit {
     if (contextMenuItem.itemId === RouteSegments.edit) {
       await this.router.navigate([RouteSegments.edit, contextMenuItem.rowData.id], { relativeTo: this.activatedRoute });
     } else if (contextMenuItem.itemId === 'delete' && contextMenuItem.rowData.id !== undefined) {
-      this.client
-        .remove(contextMenuItem.rowData.id)
-        .subscribe(() => this.getProducts(this.query));
+      await lastValueFrom(this.client.remove(contextMenuItem.rowData.id));
+      this.data = await this.getProducts(this.query);
     }
   };
 
@@ -66,12 +62,15 @@ export class ProductListComponent extends BaseListComponent implements OnInit {
     return this.activatedRoute.queryParams
       .pipe(
         tap(params => this.query.applyRouteChanges(params)),
-        switchMap(() => this.getProducts(this.query))
+        // eslint-disable-next-line @typescript-eslint/return-await
+        switchMap(async() => await this.getProducts(this.query))
       )
       .subscribe(response => this.data = response);
   }
 
-  private getProducts = (query: GetProductsQuery): Observable<PagedData<GetProductsResponseItem>> =>
-    this.client.search(query.name.value, query.code.value, query.expiresBeforeDate,
-      query.pageNumber, query.pageSize, query.sortBy, query.sortDirection);
+  private getProducts = async(query: GetProductsQuery): Promise<PagedResponseOfGetProductsResponseItem> => {
+    const result = await lastValueFrom(this.client.search(query.name.value, query.code.value, query.expiresBeforeDate(),
+      query.pageNumber, query.pageSize, query.sortBy, query.sortDirection));
+    return result;
+  };
 }
