@@ -1,37 +1,40 @@
-import { Location } from '@angular/common';
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { CommonModule, Location } from '@angular/common';
+import { Component, inject, OnInit, viewChild } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import {
   CreateOrUpdateUserCommand, IGetUserDetailsResponse,
   IValidationProblemDetails,
   LookupResponseOfGuid, LookupResponseOfUserStatusId, PermissionId, UsersClient
 } from '@api';
-import { setServerSideValidationErrors } from '@enigmatry/entry-components';
-import { Observable } from 'rxjs';
+import { EntryPermissionModule, setServerSideValidationErrors } from '@enigmatry/entry-components';
+import { FormWrapperComponent } from '@shared/form-wrapper/form-wrapper.component';
+import { NotNullPipe } from '@shared/pipes/not-null.pipe';
+import { lastValueFrom, Observable } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { UserEditGeneratedComponent } from '../generated/user-edit/user-edit-generated.component';
+import { UsersGeneratedModule } from '../generated/users-generated.module';
 
 @Component({
-  standalone: false,
+  imports: [CommonModule, FormWrapperComponent, EntryPermissionModule, UsersGeneratedModule, NotNullPipe],
   selector: 'app-user-edit',
   templateUrl: './user-edit.component.html',
   styleUrls: ['./user-edit.component.scss']
 })
 export class UserEditComponent implements OnInit {
-  @ViewChild('formComponent') formComponent: UserEditGeneratedComponent;
+  readonly formComponent = viewChild<UserEditGeneratedComponent>('formComponent');
 
   model: IGetUserDetailsResponse = {};
   $roleLookup: Observable<LookupResponseOfGuid[]>;
   $userStatusLookup: Observable<LookupResponseOfUserStatusId[]>;
   PermissionId = PermissionId;
 
-  constructor(private client: UsersClient,
-    protected activatedRoute: ActivatedRoute,
-    public location: Location) { }
+  private readonly client: UsersClient = inject(UsersClient);
+  private readonly activatedRoute: ActivatedRoute = inject(ActivatedRoute);
+  readonly location: Location = inject(Location);
 
   ngOnInit(): void {
     this.activatedRoute.params
-      .pipe(switchMap(params => this.client.get(params.id)))
+      .pipe(switchMap(params => this.client.get(params['id'])))
       .subscribe(response => {
         this.model = response;
       });
@@ -40,7 +43,7 @@ export class UserEditComponent implements OnInit {
     this.$userStatusLookup = this.client.getStatusesLookup();
   }
 
-  save(model: IGetUserDetailsResponse) {
+  readonly save = async(model: IGetUserDetailsResponse) => {
     const command = new CreateOrUpdateUserCommand({
       id: model.id,
       emailAddress: model.emailAddress,
@@ -48,16 +51,17 @@ export class UserEditComponent implements OnInit {
       userStatusId: model.userStatusId,
       fullName: model.fullName
     });
-    this.client.post(command)
-      .subscribe({
-        next: () => this.location.back(),
-        error: (error: IValidationProblemDetails) => {
-          setServerSideValidationErrors(error, this.formComponent.form);
-        }
-      });
-  }
 
-  cancel() {
+    try {
+      await lastValueFrom(this.client.post(command));
+      this.location.back();
+    } catch (error) {
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      setServerSideValidationErrors(error as IValidationProblemDetails, this.formComponent()!.form);
+    }
+  };
+
+  readonly cancel = () => {
     this.location.back();
-  }
+  };
 }
