@@ -1,16 +1,17 @@
-import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Component, inject, OnInit } from '@angular/core';
 import { GetUsersResponseItem, PermissionId, UsersClient } from '@api';
 import { PermissionService } from '@app/auth/permissions.service';
+import { EntrySearchFilterModule } from '@enigmatry/entry-components';
 import { ContextMenuItem, PagedData } from '@enigmatry/entry-components/table';
 import { BaseListComponent } from '@shared/list-component/base-list-component.model';
 import { SearchFilterPagedQuery } from '@shared/list-component/search-filter-paged-query';
 import { RouteSegments } from '@shared/model/route-segments';
 import { Observable, Subscription, switchMap, tap } from 'rxjs';
+import { UsersGeneratedModule } from '../generated/users-generated.module';
 import { GetUsersQuery } from '../models/qet-users-query.model';
 
 @Component({
-  standalone: false,
+  imports: [EntrySearchFilterModule, UsersGeneratedModule],
   selector: 'app-user-list',
   templateUrl: './user-list.component.html',
   styleUrls: ['./user-list.component.scss']
@@ -18,13 +19,9 @@ import { GetUsersQuery } from '../models/qet-users-query.model';
 export class UserListComponent extends BaseListComponent implements OnInit {
   contextMenuItems: ContextMenuItem[] = [];
   data: PagedData<GetUsersResponseItem>;
-  query = new GetUsersQuery();
-
-  constructor(private client: UsersClient, private permissionService: PermissionService,
-    protected router: Router, protected activatedRoute: ActivatedRoute
-  ) {
-    super(router, activatedRoute);
-  }
+  override query = new GetUsersQuery();
+  private readonly permissionService: PermissionService = inject(PermissionService);
+  private readonly client: UsersClient = inject(UsersClient);
 
   ngOnInit(): void {
     this.watchQueryParamsAndGetUsers();
