@@ -1,10 +1,9 @@
 //----------------------
 // <auto-generated>
-//     Generated using the NSwag toolchain v14.4.0.0 (NJsonSchema v11.3.2.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
+//     Generated using the NSwag toolchain v14.5.0.0 (NJsonSchema v11.4.0.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
 // </auto-generated>
 //----------------------
 
-/* tslint:disable */
 /* eslint-disable */
 // ReSharper disable InconsistentNaming
 
@@ -148,19 +147,19 @@ export class UsersClient implements IUsersClient {
             result200 = GetUserDetailsResponse.fromJS(resultData200);
             return _observableOf(result200);
             }));
-        } else if (status === 404) {
-            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
-            let result404: any = null;
-            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
-            result404 = ProblemDetails.fromJS(resultData404);
-            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
-            }));
         } else if (status === 400) {
             return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
             let result400: any = null;
             let resultData400 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
             result400 = ProblemDetails.fromJS(resultData400);
             return throwException("A server side error occurred.", status, _responseText, _headers, result400);
+            }));
+        } else if (status === 404) {
+            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
+            let result404: any = null;
+            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result404 = ProblemDetails.fromJS(resultData404);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
             }));
         } else if (status !== 200 && status !== 204) {
             return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
@@ -475,19 +474,19 @@ export class ProductsClient implements IProductsClient {
             result200 = ProductCreateOrUpdateResult.fromJS(resultData200);
             return _observableOf(result200);
             }));
-        } else if (status === 404) {
-            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
-            let result404: any = null;
-            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
-            result404 = ProblemDetails.fromJS(resultData404);
-            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
-            }));
         } else if (status === 400) {
             return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
             let result400: any = null;
             let resultData400 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
             result400 = ValidationProblemDetails.fromJS(resultData400);
             return throwException("A server side error occurred.", status, _responseText, _headers, result400);
+            }));
+        } else if (status === 404) {
+            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
+            let result404: any = null;
+            let resultData404 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
+            result404 = ProblemDetails.fromJS(resultData404);
+            return throwException("A server side error occurred.", status, _responseText, _headers, result404);
             }));
         } else if (status !== 200 && status !== 204) {
             return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
@@ -811,7 +810,7 @@ export class PagedResponseOfGetUsersResponseItem implements IPagedResponseOfGetU
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -844,7 +843,7 @@ export class PagedResponseOfGetUsersResponseItem implements IPagedResponseOfGetU
         if (Array.isArray(this.items)) {
             data["items"] = [];
             for (let item of this.items)
-                data["items"].push(item ? item.toJSON() : <any>undefined);
+                data["items"].push(item ? item.toJSON() : undefined as any);
         }
         data["totalCount"] = this.totalCount;
         data["pageNumber"] = this.pageNumber;
@@ -878,7 +877,7 @@ export class GetUsersResponseItem implements IGetUsersResponseItem {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -907,8 +906,8 @@ export class GetUsersResponseItem implements IGetUsersResponseItem {
         data["emailAddress"] = this.emailAddress;
         data["fullName"] = this.fullName;
         data["userStatusName"] = this.userStatusName;
-        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>undefined;
-        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>undefined;
+        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : undefined as any;
+        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : undefined as any;
         return data;
     }
 }
@@ -937,7 +936,7 @@ export class GetUserDetailsResponse implements IGetUserDetailsResponse {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -969,8 +968,8 @@ export class GetUserDetailsResponse implements IGetUserDetailsResponse {
         data["emailAddress"] = this.emailAddress;
         data["fullName"] = this.fullName;
         data["roleId"] = this.roleId;
-        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>undefined;
-        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : <any>undefined;
+        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : undefined as any;
+        data["updatedOn"] = this.updatedOn ? this.updatedOn.toISOString() : undefined as any;
         data["userStatusId"] = this.userStatusId;
         data["userStatusName"] = this.userStatusName;
         data["userStatusDescription"] = this.userStatusDescription;
@@ -1011,7 +1010,7 @@ export abstract class SmartEnumOfUserStatusIdAndInteger implements ISmartEnumOfU
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1057,7 +1056,7 @@ export class ProblemDetails implements IProblemDetails {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1119,7 +1118,7 @@ export class CreateOrUpdateUserCommand implements ICreateOrUpdateUserCommand {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1168,7 +1167,7 @@ export class LookupResponseOfGuid implements ILookupResponseOfGuid {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1208,7 +1207,7 @@ export class LookupResponseOfUserStatusId implements ILookupResponseOfUserStatus
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1253,7 +1252,7 @@ export class PagedResponseOfGetProductsResponseItem implements IPagedResponseOfG
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1286,7 +1285,7 @@ export class PagedResponseOfGetProductsResponseItem implements IPagedResponseOfG
         if (Array.isArray(this.items)) {
             data["items"] = [];
             for (let item of this.items)
-                data["items"].push(item ? item.toJSON() : <any>undefined);
+                data["items"].push(item ? item.toJSON() : undefined as any);
         }
         data["totalCount"] = this.totalCount;
         data["pageNumber"] = this.pageNumber;
@@ -1327,7 +1326,7 @@ export class GetProductsResponseItem implements IGetProductsResponseItem {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1368,7 +1367,7 @@ export class GetProductsResponseItem implements IGetProductsResponseItem {
         data["contactEmail"] = this.contactEmail;
         data["contactPhone"] = this.contactPhone;
         data["infoLink"] = this.infoLink;
-        data["expiresOn"] = this.expiresOn ? formatDate(this.expiresOn) : <any>undefined;
+        data["expiresOn"] = this.expiresOn ? formatDate(this.expiresOn) : undefined as any;
         data["freeShipping"] = this.freeShipping;
         data["hasDiscount"] = this.hasDiscount;
         data["discount"] = this.discount;
@@ -1420,7 +1419,7 @@ export class GetProductDetailsResponse implements IGetProductDetailsResponse {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1465,7 +1464,7 @@ export class GetProductDetailsResponse implements IGetProductDetailsResponse {
         data["contactPhone"] = this.contactPhone;
         data["infoLink"] = this.infoLink;
         data["additionalInfo"] = this.additionalInfo;
-        data["expiresOn"] = this.expiresOn ? formatDate(this.expiresOn) : <any>undefined;
+        data["expiresOn"] = this.expiresOn ? formatDate(this.expiresOn) : undefined as any;
         data["hasDiscount"] = this.hasDiscount;
         data["discount"] = this.discount;
         data["freeShipping"] = this.freeShipping;
@@ -1498,7 +1497,7 @@ export class IsProductCodeUniqueResponse implements IIsProductCodeUniqueResponse
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1534,7 +1533,7 @@ export class IsProductNameUniqueResponse implements IIsProductNameUniqueResponse
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1570,7 +1569,7 @@ export class ProductCreateOrUpdateResult implements IProductCreateOrUpdateResult
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1608,7 +1607,7 @@ export class HttpValidationProblemDetails extends ProblemDetails implements IHtt
         super(data);
     }
 
-    init(_data?: any) {
+    override init(_data?: any) {
         super.init(_data);
         if (_data) {
             for (var property in _data) {
@@ -1625,14 +1624,14 @@ export class HttpValidationProblemDetails extends ProblemDetails implements IHtt
         }
     }
 
-    static fromJS(data: any): HttpValidationProblemDetails {
+    static override fromJS(data: any): HttpValidationProblemDetails {
         data = typeof data === 'object' ? data : {};
         let result = new HttpValidationProblemDetails();
         result.init(data);
         return result;
     }
 
-    toJSON(data?: any) {
+    override toJSON(data?: any) {
         data = typeof data === 'object' ? data : {};
         for (var property in this) {
             if (this.hasOwnProperty(property))
@@ -1642,7 +1641,7 @@ export class HttpValidationProblemDetails extends ProblemDetails implements IHtt
             data["errors"] = {};
             for (let key in this.errors) {
                 if (this.errors.hasOwnProperty(key))
-                    (<any>data["errors"])[key] = (<any>this.errors)[key];
+                    (data["errors"] as any)[key] = (this.errors as any)[key];
             }
         }
         super.toJSON(data);
@@ -1665,7 +1664,7 @@ export class ValidationProblemDetails extends HttpValidationProblemDetails imple
         super(data);
     }
 
-    init(_data?: any) {
+    override init(_data?: any) {
         super.init(_data);
         if (_data) {
             for (var property in _data) {
@@ -1682,14 +1681,14 @@ export class ValidationProblemDetails extends HttpValidationProblemDetails imple
         }
     }
 
-    static fromJS(data: any): ValidationProblemDetails {
+    static override fromJS(data: any): ValidationProblemDetails {
         data = typeof data === 'object' ? data : {};
         let result = new ValidationProblemDetails();
         result.init(data);
         return result;
     }
 
-    toJSON(data?: any) {
+    override toJSON(data?: any) {
         data = typeof data === 'object' ? data : {};
         for (var property in this) {
             if (this.hasOwnProperty(property))
@@ -1699,7 +1698,7 @@ export class ValidationProblemDetails extends HttpValidationProblemDetails imple
             data["errors"] = {};
             for (let key in this.errors) {
                 if (this.errors.hasOwnProperty(key))
-                    (<any>data["errors"])[key] = (<any>this.errors)[key];
+                    (data["errors"] as any)[key] = (this.errors as any)[key];
             }
         }
         super.toJSON(data);
@@ -1733,7 +1732,7 @@ export class ProductCreateOrUpdateCommand implements IProductCreateOrUpdateComma
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1776,7 +1775,7 @@ export class ProductCreateOrUpdateCommand implements IProductCreateOrUpdateComma
         data["contactEmail"] = this.contactEmail;
         data["contactPhone"] = this.contactPhone;
         data["infoLink"] = this.infoLink;
-        data["expiresOn"] = this.expiresOn ? formatDate(this.expiresOn) : <any>undefined;
+        data["expiresOn"] = this.expiresOn ? formatDate(this.expiresOn) : undefined as any;
         data["freeShipping"] = this.freeShipping;
         data["hasDiscount"] = this.hasDiscount;
         data["discount"] = this.discount;
@@ -1811,7 +1810,7 @@ export class GetUserProfileResponse implements IGetUserProfileResponse {
         if (data) {
             for (var property in data) {
                 if (data.hasOwnProperty(property))
-                    (<any>this)[property] = (<any>data)[property];
+                    (this as any)[property] = (data as any)[property];
             }
         }
     }
@@ -1879,7 +1878,7 @@ function parseDateOnly(s: string) {
 }
 
 export class ApiException extends Error {
-    message: string;
+    override message: string;
     status: number;
     response: string;
     headers: { [key: string]: any; };
