@@ -1,39 +1,31 @@
 import { HttpStatusCode } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { inject, Injectable, signal } from '@angular/core';
 import { ProfileClient } from '@api';
 import { UserProfile } from '@app/auth/user-profile';
-import { BehaviorSubject, Observable, catchError, map, of, tap, throwError } from 'rxjs';
+import { lastValueFrom } from 'rxjs';
 
 @Injectable({
 	providedIn: 'root'
 })
 export class CurrentUserService {
-	private currentUserSubject = new BehaviorSubject<UserProfile | null>(null);
+	readonly currentUser = signal<UserProfile | null>(null);
+	private readonly profileClient: ProfileClient = inject(ProfileClient);
 
-	public get currentUser$(): Observable<UserProfile | null> {
-		return this.currentUserSubject.asObservable();
-	}
-
-	public get currentUser(): UserProfile | null {
-		return this.currentUserSubject.value;
-	}
-
-	constructor(private profileClient: ProfileClient) { }
-
-	loadUser(): Observable<UserProfile | null> {
-		if (this.currentUserSubject.value) {
-			return this.currentUserSubject.asObservable();
+	readonly loadUser = async() => {
+		if (this.currentUser()) {
+			return;
 		}
-		return this.profileClient.getProfile()
-			.pipe(
-				map(response => UserProfile.fromResponse(response)),
-				tap(profile => this.currentUserSubject.next(profile)),
-				catchError(error => {
-					if (error.status === HttpStatusCode.NotFound) {
-						return of(null);
-					}
-					return throwError(() => error);
-				})
-			);
-	}
+
+		try {
+			const response = await lastValueFrom(this.profileClient.getProfile());
+			const profile = UserProfile.fromResponse(response);
+			this.currentUser.set(profile);
+		} catch (error) {
+			if (error as Record<string, unknown>['status'] === HttpStatusCode.NotFound) {
+				this.currentUser.set(null);
+				return;
+			}
+			throw error;
+		}
+	};
 }
