@@ -19,90 +19,90 @@ import { BehaviorSubject, of, Subject, Subscription } from 'rxjs';
 import { map, throttleTime } from 'rxjs/operators';
 
 
-@Component({
-  standalone: false,
-  selector: 'app-g-product-edit',
-  templateUrl: './product-edit-generated.component.html'
-})
-export class ProductEditGeneratedComponent implements OnInit, OnDestroy {
+    @Component({
+    standalone: false,
+    selector: 'app-g-product-edit',
+    templateUrl: './product-edit-generated.component.html'
+    })
+    export class ProductEditGeneratedComponent implements OnInit, OnDestroy {
 
-  @Input() model: IGetProductDetailsResponse = {} as IGetProductDetailsResponse;
-  @Input() set isReadonly(value: boolean) {
+    @Input() model: IGetProductDetailsResponse = {} as IGetProductDetailsResponse;
+    @Input() set isReadonly(value: boolean) {
     this._isReadonly = value;
     this.fields = this.initializeFields();
-  }
-  get isReadonly() {
+    }
+    get isReadonly() {
     return this._isReadonly;
-  }
+    }
 
-  @Input() saveButtonText: string = 'Save';
-  @Input() cancelButtonText: string = 'Cancel';
-  @Input() saveButtonDisabled: boolean = false;
-  @Input() formButtonsTemplate: TemplateRef<any> | null | undefined;
+    @Input() saveButtonText: string = 'Save';
+    @Input() cancelButtonText: string = 'Cancel';
+    @Input() saveButtonDisabled: boolean = false;
+    @Input() formButtonsTemplate: TemplateRef<any> | null | undefined;
 
-  @Input() fieldsHideExpressions: IFieldExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
-  @Input() fieldsDisableExpressions: IFieldExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
-  @Input() fieldsRequiredExpressions: IFieldExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
-  @Input() fieldsPropertyExpressions: IFieldPropertyExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
-  @Input() fieldsLabelExpressions: IFieldPropertyExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
+    @Input() fieldsHideExpressions: IFieldExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
+    @Input() fieldsDisableExpressions: IFieldExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
+    @Input() fieldsRequiredExpressions: IFieldExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
+    @Input() fieldsPropertyExpressions: IFieldPropertyExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
+    @Input() fieldsLabelExpressions: IFieldPropertyExpressionDictionary<IGetProductDetailsResponse> | undefined = undefined;
 
-  @Output() save = new EventEmitter<IGetProductDetailsResponse>();
-  @Output() cancel = new EventEmitter<void>();
-  @Output() buttonClick = new EventEmitter<string>();
+    @Output() save = new EventEmitter<IGetProductDetailsResponse>();
+    @Output() cancel = new EventEmitter<void>();
+    @Output() buttonClick = new EventEmitter<string>();
 
             @Input() typeOptions: any[] = [{ value: 0, displayName: $localize `:@@enum.product-type.food:Food` }, { value: 1, displayName: $localize `:@@enum.product-type.drink:Drink` }, { value: 2, displayName: $localize `:@@enum.product-type.book:Book` }, { value: 3, displayName: $localize `:@@enum.product-type.car:Car` }];
             @Input() typeOptionsConfiguration: SelectConfiguration = { valueProperty: 'value', labelProperty: 'displayName', sortProperty: 'displayName' };
 
-  _isReadonly: boolean;
-  form = new FormGroup({});
-  fields: FormlyFieldConfig[] = [];
-  private _submitClicks = new Subject<void>();
-  private _submitClicksSubscription: Subscription;
+    _isReadonly: boolean;
+    form = new FormGroup({});
+    fields: FormlyFieldConfig[] = [];
+    private _submitClicks = new Subject<void>();
+    private _submitClicksSubscription: Subscription;
 
-  constructor(
+    constructor(
     @Inject(LOCALE_ID) private _localeId: string,
     @Optional() @Inject(ENTRY_FIELD_TYPE_RESOLVER) private _fieldTypeResolver: FieldTypeResolver) { }
 
-  ngOnInit(): void {
+    ngOnInit(): void {
     this.fields = this.initializeFields();
     this._submitClicksSubscription = this._submitClicks
-        .pipe(throttleTime(500))
-        .subscribe(() => this.save.emit(this.model));
-  }
-
-  ngOnDestroy(): void {
-      this._submitClicksSubscription.unsubscribe();
-  }
-
-  onSubmit() {
-    if (this.form.valid) {
-      this._submitClicks.next();
+    .pipe(throttleTime(500))
+    .subscribe(() => this.save.emit(this.model));
     }
-  }
 
-  resolveFieldType = (type: string, isControlReadonly: boolean): string =>
+    ngOnDestroy(): void {
+    this._submitClicksSubscription.unsubscribe();
+    }
+
+    onSubmit() {
+    if (this.form.valid) {
+    this._submitClicks.next();
+    }
+    }
+
+    resolveFieldType = (type: string, isControlReadonly: boolean): string =>
     this._fieldTypeResolver ? this._fieldTypeResolver(type, this.isReadonly || isControlReadonly) : type;
 
-  initializeFields(): FormlyFieldConfig[] {
+    initializeFields(): FormlyFieldConfig[] {
     return [
         {
         key: 'name',
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-name-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.name ?? false,
+        hideExpression: this.fieldsHideExpressions?.['name'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.name ? this.fieldsDisableExpressions.name(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.name ? this.fieldsRequiredExpressions.name(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.name ? this.fieldsLabelExpressions.name(model) : $localize `:@@products.product-edit.name.label:Name`),
-        'model.name': (model) => (this.fieldsPropertyExpressions?.name ? this.fieldsPropertyExpressions.name(model) : model.name),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['name'] ? this.fieldsDisableExpressions['name'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['name'] ? this.fieldsRequiredExpressions['name'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['name'] ? this.fieldsLabelExpressions['name'](model) : $localize `:@@products.product-edit.name.label:Name`),
+        'model.name': (model) => (this.fieldsPropertyExpressions?.['name'] ? this.fieldsPropertyExpressions['name'](model) : model.name),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.name.label:Name`,
         placeholder: $localize `:@@products.product-edit.name.placeholder:Unique product name`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 minLength: 5,
 maxLength: 50,
@@ -116,19 +116,19 @@ asyncValidators: { validation: [ 'productNameIsUnique' ] },
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-code-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.code ?? false,
+        hideExpression: this.fieldsHideExpressions?.['code'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.code ? this.fieldsDisableExpressions.code(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.code ? this.fieldsRequiredExpressions.code(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.code ? this.fieldsLabelExpressions.code(model) : $localize `:@@products.product-edit.code.label:Code`),
-        'model.code': (model) => (this.fieldsPropertyExpressions?.code ? this.fieldsPropertyExpressions.code(model) : model.code),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['code'] ? this.fieldsDisableExpressions['code'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['code'] ? this.fieldsRequiredExpressions['code'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['code'] ? this.fieldsLabelExpressions['code'](model) : $localize `:@@products.product-edit.code.label:Code`),
+        'model.code': (model) => (this.fieldsPropertyExpressions?.['code'] ? this.fieldsPropertyExpressions['code'](model) : model.code),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.code.label:Code`,
         placeholder: $localize `:@@products.product-edit.code.placeholder:Unique product code identifier`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 pattern: /^[A-Z]{4}[1-9]{8}$/mu,
 
@@ -146,12 +146,12 @@ asyncValidators: { validation: [ 'productCodeIsUnique' ] },
         type: this.resolveFieldType('autocomplete', false),
         focus: false,
 className: `entry-type-field entry-autocomplete`,
-        hideExpression: this.fieldsHideExpressions?.type ?? false,
+        hideExpression: this.fieldsHideExpressions?.['type'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.type ? this.fieldsDisableExpressions.type(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.type ? this.fieldsRequiredExpressions.type(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.type ? this.fieldsLabelExpressions.type(model) : $localize `:@@products.product-edit.type.label:Type`),
-        'model.type': (model) => (this.fieldsPropertyExpressions?.type ? this.fieldsPropertyExpressions.type(model) : model.type),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['type'] ? this.fieldsDisableExpressions['type'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['type'] ? this.fieldsRequiredExpressions['type'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['type'] ? this.fieldsLabelExpressions['type'](model) : $localize `:@@products.product-edit.type.label:Type`),
+        'model.type': (model) => (this.fieldsPropertyExpressions?.['type'] ? this.fieldsPropertyExpressions['type'](model) : model.type),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.type.label:Type`,
@@ -161,7 +161,7 @@ className: `entry-type-field entry-autocomplete`,
             valueProp: this.typeOptionsConfiguration.valueProperty,
             labelProp: this.typeOptionsConfiguration.labelProperty,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 
             typeFormatDef: undefined
@@ -172,12 +172,12 @@ className: `entry-type-field entry-autocomplete`,
         type: this.resolveFieldType('textarea', false),
         focus: false,
 className: `entry-description-field entry-textarea`,
-        hideExpression: this.fieldsHideExpressions?.description ?? false,
+        hideExpression: this.fieldsHideExpressions?.['description'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.description ? this.fieldsDisableExpressions.description(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.description ? this.fieldsRequiredExpressions.description(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.description ? this.fieldsLabelExpressions.description(model) : $localize `:@@products.product-edit.description.label:Description`),
-        'model.description': (model) => (this.fieldsPropertyExpressions?.description ? this.fieldsPropertyExpressions.description(model) : model.description),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['description'] ? this.fieldsDisableExpressions['description'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['description'] ? this.fieldsRequiredExpressions['description'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['description'] ? this.fieldsLabelExpressions['description'](model) : $localize `:@@products.product-edit.description.label:Description`),
+        'model.description': (model) => (this.fieldsPropertyExpressions?.['description'] ? this.fieldsPropertyExpressions['description'](model) : model.description),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.description.label:Description`,
@@ -189,7 +189,7 @@ className: `entry-description-field entry-textarea`,
             autosizeMinRows: 0,
             autosizeMaxRows: 0,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: undefined
         },
         },
@@ -198,19 +198,19 @@ className: `entry-description-field entry-textarea`,
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-price-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.price ?? false,
+        hideExpression: this.fieldsHideExpressions?.['price'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.price ? this.fieldsDisableExpressions.price(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.price ? this.fieldsRequiredExpressions.price(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.price ? this.fieldsLabelExpressions.price(model) : $localize `:@@products.price:Price per unit`),
-        'model.price': (model) => (this.fieldsPropertyExpressions?.price ? this.fieldsPropertyExpressions.price(model) : model.price),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['price'] ? this.fieldsDisableExpressions['price'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['price'] ? this.fieldsRequiredExpressions['price'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['price'] ? this.fieldsLabelExpressions['price'](model) : $localize `:@@products.price:Price per unit`),
+        'model.price': (model) => (this.fieldsPropertyExpressions?.['price'] ? this.fieldsPropertyExpressions['price'](model) : model.price),
         },
         templateOptions: {
         label: $localize `:@@products.price:Price per unit`,
         placeholder: $localize `:@@products.price:Price per unit`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 type: 'number',
 min: 0.99 + 0.1,
@@ -224,19 +224,19 @@ max: 999.99,
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-amount-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.amount ?? false,
+        hideExpression: this.fieldsHideExpressions?.['amount'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.amount ? this.fieldsDisableExpressions.amount(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.amount ? this.fieldsRequiredExpressions.amount(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.amount ? this.fieldsLabelExpressions.amount(model) : $localize `:@@products.amount:Units`),
-        'model.amount': (model) => (this.fieldsPropertyExpressions?.amount ? this.fieldsPropertyExpressions.amount(model) : model.amount),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['amount'] ? this.fieldsDisableExpressions['amount'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['amount'] ? this.fieldsRequiredExpressions['amount'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['amount'] ? this.fieldsLabelExpressions['amount'](model) : $localize `:@@products.amount:Units`),
+        'model.amount': (model) => (this.fieldsPropertyExpressions?.['amount'] ? this.fieldsPropertyExpressions['amount'](model) : model.amount),
         },
         templateOptions: {
         label: $localize `:@@products.amount:Units`,
         placeholder: $localize `:@@products.amount:Units`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 type: 'number',
 min: 0 + 1,
@@ -249,19 +249,19 @@ min: 0 + 1,
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-contact-email-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.contactEmail ?? false,
+        hideExpression: this.fieldsHideExpressions?.['contactEmail'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.contactEmail ? this.fieldsDisableExpressions.contactEmail(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.contactEmail ? this.fieldsRequiredExpressions.contactEmail(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.contactEmail ? this.fieldsLabelExpressions.contactEmail(model) : $localize `:@@products.product-edit.contact-email.label:Contact email`),
-        'model.contactEmail': (model) => (this.fieldsPropertyExpressions?.contactEmail ? this.fieldsPropertyExpressions.contactEmail(model) : model.contactEmail),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['contactEmail'] ? this.fieldsDisableExpressions['contactEmail'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['contactEmail'] ? this.fieldsRequiredExpressions['contactEmail'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['contactEmail'] ? this.fieldsLabelExpressions['contactEmail'](model) : $localize `:@@products.product-edit.contact-email.label:Contact email`),
+        'model.contactEmail': (model) => (this.fieldsPropertyExpressions?.['contactEmail'] ? this.fieldsPropertyExpressions['contactEmail'](model) : model.contactEmail),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.contact-email.label:Contact email`,
         placeholder: $localize `:@@products.product-edit.contact-email.placeholder:Contact person email address`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 pattern: /^\w+([-+.']\w+)*@\w+([-.]\w+)*\.\w+([-.]\w+)*$/,
 
@@ -278,19 +278,19 @@ pattern: (err, field) => $localize `:@@validators.pattern.emailAddress:Invalid e
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-contact-phone-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.contactPhone ?? false,
+        hideExpression: this.fieldsHideExpressions?.['contactPhone'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.contactPhone ? this.fieldsDisableExpressions.contactPhone(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.contactPhone ? this.fieldsRequiredExpressions.contactPhone(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.contactPhone ? this.fieldsLabelExpressions.contactPhone(model) : $localize `:@@products.product-edit.contact-phone.label:Contact phone`),
-        'model.contactPhone': (model) => (this.fieldsPropertyExpressions?.contactPhone ? this.fieldsPropertyExpressions.contactPhone(model) : model.contactPhone),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['contactPhone'] ? this.fieldsDisableExpressions['contactPhone'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['contactPhone'] ? this.fieldsRequiredExpressions['contactPhone'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['contactPhone'] ? this.fieldsLabelExpressions['contactPhone'](model) : $localize `:@@products.product-edit.contact-phone.label:Contact phone`),
+        'model.contactPhone': (model) => (this.fieldsPropertyExpressions?.['contactPhone'] ? this.fieldsPropertyExpressions['contactPhone'](model) : model.contactPhone),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.contact-phone.label:Contact phone`,
         placeholder: $localize `:@@products.product-edit.contact-phone.placeholder:Contact person phone number`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 pattern: /^s*(?:\+?(\d{1,3}))?[-. (]*(\d{3})[-. )]*(\d{3})[-. ]*(\d{4})(?: *x(\d+))?\s*$/mu,
 
@@ -302,19 +302,19 @@ pattern: /^s*(?:\+?(\d{1,3}))?[-. (]*(\d{3})[-. )]*(\d{3})[-. ]*(\d{4})(?: *x(\d
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-info-link-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.infoLink ?? false,
+        hideExpression: this.fieldsHideExpressions?.['infoLink'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.infoLink ? this.fieldsDisableExpressions.infoLink(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.infoLink ? this.fieldsRequiredExpressions.infoLink(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.infoLink ? this.fieldsLabelExpressions.infoLink(model) : $localize `:@@products.product-edit.info-link.label:Homepage`),
-        'model.infoLink': (model) => (this.fieldsPropertyExpressions?.infoLink ? this.fieldsPropertyExpressions.infoLink(model) : model.infoLink),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['infoLink'] ? this.fieldsDisableExpressions['infoLink'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['infoLink'] ? this.fieldsRequiredExpressions['infoLink'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['infoLink'] ? this.fieldsLabelExpressions['infoLink'](model) : $localize `:@@products.product-edit.info-link.label:Homepage`),
+        'model.infoLink': (model) => (this.fieldsPropertyExpressions?.['infoLink'] ? this.fieldsPropertyExpressions['infoLink'](model) : model.infoLink),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.info-link.label:Homepage`,
         placeholder: $localize `:@@products.product-edit.info-link.placeholder:Link to product homepage`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             pattern: /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)/u,
 
             typeFormatDef: undefined
@@ -325,19 +325,19 @@ className: `entry-info-link-field entry-input`,
         type: this.resolveFieldType('datepicker', false),
         focus: false,
 className: `entry-expires-on-field entry-datepicker`,
-        hideExpression: this.fieldsHideExpressions?.expiresOn ?? false,
+        hideExpression: this.fieldsHideExpressions?.['expiresOn'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.expiresOn ? this.fieldsDisableExpressions.expiresOn(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.expiresOn ? this.fieldsRequiredExpressions.expiresOn(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.expiresOn ? this.fieldsLabelExpressions.expiresOn(model) : $localize `:@@products.product-edit.expires-on.label:Expires on`),
-        'model.expiresOn': (model) => (this.fieldsPropertyExpressions?.expiresOn ? this.fieldsPropertyExpressions.expiresOn(model) : model.expiresOn),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['expiresOn'] ? this.fieldsDisableExpressions['expiresOn'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['expiresOn'] ? this.fieldsRequiredExpressions['expiresOn'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['expiresOn'] ? this.fieldsLabelExpressions['expiresOn'](model) : $localize `:@@products.product-edit.expires-on.label:Expires on`),
+        'model.expiresOn': (model) => (this.fieldsPropertyExpressions?.['expiresOn'] ? this.fieldsPropertyExpressions['expiresOn'](model) : model.expiresOn),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.expires-on.label:Expires on`,
         placeholder: $localize `:@@products.product-edit.expires-on.placeholder:Product expiration date, if any`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: { name: 'date' }
         },
 modelOptions: { updateOn: 'blur' },
@@ -347,19 +347,19 @@ modelOptions: { updateOn: 'blur' },
         type: this.resolveFieldType('checkbox', false),
         focus: false,
 className: `entry-free-shipping-field entry-checkbox`,
-        hideExpression: this.fieldsHideExpressions?.freeShipping ?? false,
+        hideExpression: this.fieldsHideExpressions?.['freeShipping'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.freeShipping ? this.fieldsDisableExpressions.freeShipping(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.freeShipping ? this.fieldsRequiredExpressions.freeShipping(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.freeShipping ? this.fieldsLabelExpressions.freeShipping(model) : $localize `:@@products.product-edit.free-shipping.label:Free shipping`),
-        'model.freeShipping': (model) => (this.fieldsPropertyExpressions?.freeShipping ? this.fieldsPropertyExpressions.freeShipping(model) : model.freeShipping),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['freeShipping'] ? this.fieldsDisableExpressions['freeShipping'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['freeShipping'] ? this.fieldsRequiredExpressions['freeShipping'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['freeShipping'] ? this.fieldsLabelExpressions['freeShipping'](model) : $localize `:@@products.product-edit.free-shipping.label:Free shipping`),
+        'model.freeShipping': (model) => (this.fieldsPropertyExpressions?.['freeShipping'] ? this.fieldsPropertyExpressions['freeShipping'](model) : model.freeShipping),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.free-shipping.label:Free shipping`,
         placeholder: $localize `:@@products.product-edit.free-shipping.label:Free shipping`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: { name: 'boolean' }
         },
         },
@@ -368,19 +368,19 @@ className: `entry-free-shipping-field entry-checkbox`,
         type: this.resolveFieldType('checkbox', false),
         focus: false,
 className: `entry-has-discount-field entry-checkbox`,
-        hideExpression: this.fieldsHideExpressions?.hasDiscount ?? false,
+        hideExpression: this.fieldsHideExpressions?.['hasDiscount'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.hasDiscount ? this.fieldsDisableExpressions.hasDiscount(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.hasDiscount ? this.fieldsRequiredExpressions.hasDiscount(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.hasDiscount ? this.fieldsLabelExpressions.hasDiscount(model) : $localize `:@@products.product-edit.has-discount.label:Has discount`),
-        'model.hasDiscount': (model) => (this.fieldsPropertyExpressions?.hasDiscount ? this.fieldsPropertyExpressions.hasDiscount(model) : model.hasDiscount),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['hasDiscount'] ? this.fieldsDisableExpressions['hasDiscount'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['hasDiscount'] ? this.fieldsRequiredExpressions['hasDiscount'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['hasDiscount'] ? this.fieldsLabelExpressions['hasDiscount'](model) : $localize `:@@products.product-edit.has-discount.label:Has discount`),
+        'model.hasDiscount': (model) => (this.fieldsPropertyExpressions?.['hasDiscount'] ? this.fieldsPropertyExpressions['hasDiscount'](model) : model.hasDiscount),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.has-discount.label:Has discount`,
         placeholder: $localize `:@@products.product-edit.has-discount.label:Has discount`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: { name: 'boolean' }
         },
         },
@@ -389,19 +389,19 @@ className: `entry-has-discount-field entry-checkbox`,
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-discount-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.discount ?? false,
+        hideExpression: this.fieldsHideExpressions?.['discount'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.discount ? this.fieldsDisableExpressions.discount(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.discount ? this.fieldsRequiredExpressions.discount(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.discount ? this.fieldsLabelExpressions.discount(model) : $localize `:@@products.product-edit.discount.label:Discount`),
-        'model.discount': (model) => (this.fieldsPropertyExpressions?.discount ? this.fieldsPropertyExpressions.discount(model) : model.discount),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['discount'] ? this.fieldsDisableExpressions['discount'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['discount'] ? this.fieldsRequiredExpressions['discount'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['discount'] ? this.fieldsLabelExpressions['discount'](model) : $localize `:@@products.product-edit.discount.label:Discount`),
+        'model.discount': (model) => (this.fieldsPropertyExpressions?.['discount'] ? this.fieldsPropertyExpressions['discount'](model) : model.discount),
         },
         templateOptions: {
         label: $localize `:@@products.product-edit.discount.label:Discount`,
         placeholder: $localize `:@@products.product-edit.discount.label:Discount`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             type: 'number',
 min: 0,
 max: 100,
@@ -414,12 +414,12 @@ max: 100,
         type: this.resolveFieldType('button', false),
         focus: false,
 className: `entry-reset-form-btn-field entry-button`,
-        hideExpression: this.fieldsHideExpressions?.resetFormBtn ?? false,
+        hideExpression: this.fieldsHideExpressions?.['resetFormBtn'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.resetFormBtn ? this.fieldsDisableExpressions.resetFormBtn(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.resetFormBtn ? this.fieldsRequiredExpressions.resetFormBtn(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.resetFormBtn ? this.fieldsLabelExpressions.resetFormBtn(model) : ``),
-        'model.resetFormBtn': (model) => (this.fieldsPropertyExpressions?.resetFormBtn ? this.fieldsPropertyExpressions.resetFormBtn(model) : model.resetFormBtn),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['resetFormBtn'] ? this.fieldsDisableExpressions['resetFormBtn'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['resetFormBtn'] ? this.fieldsRequiredExpressions['resetFormBtn'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['resetFormBtn'] ? this.fieldsLabelExpressions['resetFormBtn'](model) : ``),
+        'model.resetFormBtn': (model) => (this.fieldsPropertyExpressions?.['resetFormBtn'] ? this.fieldsPropertyExpressions['resetFormBtn'](model) : model.resetFormBtn),
         },
         templateOptions: {
         label: ``,
@@ -428,15 +428,15 @@ className: `entry-reset-form-btn-field entry-button`,
             text: $localize `:@@products.product-edit.reset-form-btn.text:Reset`,
             onClick: ($event: any) => this.buttonClick.emit('resetFormBtn'),
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
         },
         },
             { key: 'id' },
     ];
-}
+    }
 
-  private applyOptionally<T>(value: T, apply: boolean): T | undefined {
+    private applyOptionally<T>(value: T, apply: boolean): T | undefined {
     return apply ? value : undefined;
-  }
+    }
 
-}
+    }
