@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-magic-numbers */
+import { computed } from '@angular/core';
 import { SelectOption, SelectSearchFilter, TextSearchFilter } from '@enigmatry/entry-components/search-filter';
 import { SortDirection } from '@enigmatry/entry-components/table';
 import { SearchFilterPagedQuery } from '@shared/list-component/search-filter-paged-query';
@@ -36,9 +37,9 @@ export class GetProductsQuery extends SearchFilterPagedQuery {
     this.filters = [this.name, this.code, this.expiresInMonths];
   }
 
-  get expiresBeforeDate(): Date | undefined {
+  readonly expiresBeforeDate = computed(() => {
     const expiresInMonths = this.expiresInMonths.value;
     const today = new Date(new Date().getFullYear(), new Date().getMonth(), new Date().getDate());
     return expiresInMonths ? addMonths(today, expiresInMonths) : undefined;
-  }
+  });
 }
