@@ -19,36 +19,36 @@ import { BehaviorSubject, of, Subject, Subscription } from 'rxjs';
 import { map, throttleTime } from 'rxjs/operators';
 
 
-@Component({
-  standalone: false,
-  selector: 'app-g-user-edit',
-  templateUrl: './user-edit-generated.component.html'
-})
-export class UserEditGeneratedComponent implements OnInit, OnDestroy {
+    @Component({
+    standalone: false,
+    selector: 'app-g-user-edit',
+    templateUrl: './user-edit-generated.component.html'
+    })
+    export class UserEditGeneratedComponent implements OnInit, OnDestroy {
 
-  @Input() model: IGetUserDetailsResponse = {} as IGetUserDetailsResponse;
-  @Input() set isReadonly(value: boolean) {
+    @Input() model: IGetUserDetailsResponse = {} as IGetUserDetailsResponse;
+    @Input() set isReadonly(value: boolean) {
     this._isReadonly = value;
     this.fields = this.initializeFields();
-  }
-  get isReadonly() {
+    }
+    get isReadonly() {
     return this._isReadonly;
-  }
+    }
 
-  @Input() saveButtonText: string = 'Save';
-  @Input() cancelButtonText: string = 'Cancel';
-  @Input() saveButtonDisabled: boolean = false;
-  @Input() formButtonsTemplate: TemplateRef<any> | null | undefined;
+    @Input() saveButtonText: string = 'Save';
+    @Input() cancelButtonText: string = 'Cancel';
+    @Input() saveButtonDisabled: boolean = false;
+    @Input() formButtonsTemplate: TemplateRef<any> | null | undefined;
 
-  @Input() fieldsHideExpressions: IFieldExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
-  @Input() fieldsDisableExpressions: IFieldExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
-  @Input() fieldsRequiredExpressions: IFieldExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
-  @Input() fieldsPropertyExpressions: IFieldPropertyExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
-  @Input() fieldsLabelExpressions: IFieldPropertyExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
+    @Input() fieldsHideExpressions: IFieldExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
+    @Input() fieldsDisableExpressions: IFieldExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
+    @Input() fieldsRequiredExpressions: IFieldExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
+    @Input() fieldsPropertyExpressions: IFieldPropertyExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
+    @Input() fieldsLabelExpressions: IFieldPropertyExpressionDictionary<IGetUserDetailsResponse> | undefined = undefined;
 
-  @Output() save = new EventEmitter<IGetUserDetailsResponse>();
-  @Output() cancel = new EventEmitter<void>();
-  @Output() buttonClick = new EventEmitter<string>();
+    @Output() save = new EventEmitter<IGetUserDetailsResponse>();
+    @Output() cancel = new EventEmitter<void>();
+    @Output() buttonClick = new EventEmitter<string>();
 
             private roleIdOptions$ = new BehaviorSubject<any[]>([]);
             @Input()
@@ -61,37 +61,37 @@ export class UserEditGeneratedComponent implements OnInit, OnDestroy {
             set userStatusIdOptions(value: any[]) { this.userStatusIdOptions$.next(value); }
             @Input() userStatusIdOptionsConfiguration: SelectConfiguration = {};
 
-  _isReadonly: boolean;
-  form = new FormGroup({});
-  fields: FormlyFieldConfig[] = [];
-  private _submitClicks = new Subject<void>();
-  private _submitClicksSubscription: Subscription;
+    _isReadonly: boolean;
+    form = new FormGroup({});
+    fields: FormlyFieldConfig[] = [];
+    private _submitClicks = new Subject<void>();
+    private _submitClicksSubscription: Subscription;
 
-  constructor(
+    constructor(
     @Inject(LOCALE_ID) private _localeId: string,
     @Optional() @Inject(ENTRY_FIELD_TYPE_RESOLVER) private _fieldTypeResolver: FieldTypeResolver) { }
 
-  ngOnInit(): void {
+    ngOnInit(): void {
     this.fields = this.initializeFields();
     this._submitClicksSubscription = this._submitClicks
-        .pipe(throttleTime(500))
-        .subscribe(() => this.save.emit(this.model));
-  }
-
-  ngOnDestroy(): void {
-      this._submitClicksSubscription.unsubscribe();
-  }
-
-  onSubmit() {
-    if (this.form.valid) {
-      this._submitClicks.next();
+    .pipe(throttleTime(500))
+    .subscribe(() => this.save.emit(this.model));
     }
-  }
 
-  resolveFieldType = (type: string, isControlReadonly: boolean): string =>
+    ngOnDestroy(): void {
+    this._submitClicksSubscription.unsubscribe();
+    }
+
+    onSubmit() {
+    if (this.form.valid) {
+    this._submitClicks.next();
+    }
+    }
+
+    resolveFieldType = (type: string, isControlReadonly: boolean): string =>
     this._fieldTypeResolver ? this._fieldTypeResolver(type, this.isReadonly || isControlReadonly) : type;
 
-  initializeFields(): FormlyFieldConfig[] {
+    initializeFields(): FormlyFieldConfig[] {
     return [
         {
         type: this.resolveFieldType('', false),
@@ -108,19 +108,19 @@ fieldGroupClassName: `entry-field-group`,
         type: this.resolveFieldType('input', true),
         focus: false,
 className: `entry-email-address-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.emailAddress ?? false,
+        hideExpression: this.fieldsHideExpressions?.['emailAddress'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.emailAddress ? this.fieldsDisableExpressions.emailAddress(model) : true)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.emailAddress ? this.fieldsRequiredExpressions.emailAddress(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.emailAddress ? this.fieldsLabelExpressions.emailAddress(model) : $localize `:@@users.user-edit.email-address.label:Email address`),
-        'model.emailAddress': (model) => (this.fieldsPropertyExpressions?.emailAddress ? this.fieldsPropertyExpressions.emailAddress(model) : model.emailAddress),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['emailAddress'] ? this.fieldsDisableExpressions['emailAddress'](model) : true)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['emailAddress'] ? this.fieldsRequiredExpressions['emailAddress'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['emailAddress'] ? this.fieldsLabelExpressions['emailAddress'](model) : $localize `:@@users.user-edit.email-address.label:Email address`),
+        'model.emailAddress': (model) => (this.fieldsPropertyExpressions?.['emailAddress'] ? this.fieldsPropertyExpressions['emailAddress'](model) : model.emailAddress),
         },
         templateOptions: {
         label: $localize `:@@users.user-edit.email-address.label:Email address`,
         placeholder: $localize `:@@users.user-edit.email-address.label:Email address`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: undefined
         },
         },
@@ -129,19 +129,19 @@ className: `entry-email-address-field entry-input`,
         type: this.resolveFieldType('input', false),
         focus: false,
 className: `entry-full-name-field entry-input`,
-        hideExpression: this.fieldsHideExpressions?.fullName ?? false,
+        hideExpression: this.fieldsHideExpressions?.['fullName'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.fullName ? this.fieldsDisableExpressions.fullName(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.fullName ? this.fieldsRequiredExpressions.fullName(model) : true),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.fullName ? this.fieldsLabelExpressions.fullName(model) : $localize `:@@users.user-edit.full-name.label:Full name`),
-        'model.fullName': (model) => (this.fieldsPropertyExpressions?.fullName ? this.fieldsPropertyExpressions.fullName(model) : model.fullName),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['fullName'] ? this.fieldsDisableExpressions['fullName'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['fullName'] ? this.fieldsRequiredExpressions['fullName'](model) : true),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['fullName'] ? this.fieldsLabelExpressions['fullName'](model) : $localize `:@@users.user-edit.full-name.label:Full name`),
+        'model.fullName': (model) => (this.fieldsPropertyExpressions?.['fullName'] ? this.fieldsPropertyExpressions['fullName'](model) : model.fullName),
         },
         templateOptions: {
         label: $localize `:@@users.user-edit.full-name.label:Full name`,
         placeholder: $localize `:@@users.user-edit.full-name.label:Full name`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             required: true,
 maxLength: 200,
 
@@ -153,12 +153,12 @@ maxLength: 200,
         type: this.resolveFieldType('select', false),
         focus: false,
 className: `entry-role-id-field entry-select`,
-        hideExpression: this.fieldsHideExpressions?.roleId ?? false,
+        hideExpression: this.fieldsHideExpressions?.['roleId'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.roleId ? this.fieldsDisableExpressions.roleId(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.roleId ? this.fieldsRequiredExpressions.roleId(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.roleId ? this.fieldsLabelExpressions.roleId(model) : $localize `:@@users.user-edit.role-id.label:Role`),
-        'model.roleId': (model) => (this.fieldsPropertyExpressions?.roleId ? this.fieldsPropertyExpressions.roleId(model) : model.roleId),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['roleId'] ? this.fieldsDisableExpressions['roleId'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['roleId'] ? this.fieldsRequiredExpressions['roleId'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['roleId'] ? this.fieldsLabelExpressions['roleId'](model) : $localize `:@@users.user-edit.role-id.label:Role`),
+        'model.roleId': (model) => (this.fieldsPropertyExpressions?.['roleId'] ? this.fieldsPropertyExpressions['roleId'](model) : model.roleId),
         },
         templateOptions: {
         label: $localize `:@@users.user-edit.role-id.label:Role`,
@@ -168,7 +168,7 @@ className: `entry-role-id-field entry-select`,
             valueProp: this.roleIdOptionsConfiguration.valueProperty,
             labelProp: this.roleIdOptionsConfiguration.labelProperty,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: undefined
         },
         },
@@ -177,12 +177,12 @@ className: `entry-role-id-field entry-select`,
         type: this.resolveFieldType('select', false),
         focus: false,
 className: `entry-user-status-id-field entry-select`,
-        hideExpression: this.fieldsHideExpressions?.userStatusId ?? false,
+        hideExpression: this.fieldsHideExpressions?.['userStatusId'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.userStatusId ? this.fieldsDisableExpressions.userStatusId(model) : false)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.userStatusId ? this.fieldsRequiredExpressions.userStatusId(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.userStatusId ? this.fieldsLabelExpressions.userStatusId(model) : $localize `:@@users.user-edit.user-status-id.label:Status`),
-        'model.userStatusId': (model) => (this.fieldsPropertyExpressions?.userStatusId ? this.fieldsPropertyExpressions.userStatusId(model) : model.userStatusId),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['userStatusId'] ? this.fieldsDisableExpressions['userStatusId'](model) : false)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['userStatusId'] ? this.fieldsRequiredExpressions['userStatusId'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['userStatusId'] ? this.fieldsLabelExpressions['userStatusId'](model) : $localize `:@@users.user-edit.user-status-id.label:Status`),
+        'model.userStatusId': (model) => (this.fieldsPropertyExpressions?.['userStatusId'] ? this.fieldsPropertyExpressions['userStatusId'](model) : model.userStatusId),
         },
         templateOptions: {
         label: $localize `:@@users.user-edit.user-status-id.label:Status`,
@@ -192,7 +192,7 @@ className: `entry-user-status-id-field entry-select`,
             valueProp: this.userStatusIdOptionsConfiguration.valueProperty,
             labelProp: this.userStatusIdOptionsConfiguration.labelProperty,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: undefined
         },
         },
@@ -211,19 +211,19 @@ fieldGroupClassName: `entry-field-group`,
         type: this.resolveFieldType('datepicker', true),
         focus: false,
 className: `entry-created-on-field entry-datepicker`,
-        hideExpression: this.fieldsHideExpressions?.createdOn ?? false,
+        hideExpression: this.fieldsHideExpressions?.['createdOn'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.createdOn ? this.fieldsDisableExpressions.createdOn(model) : true)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.createdOn ? this.fieldsRequiredExpressions.createdOn(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.createdOn ? this.fieldsLabelExpressions.createdOn(model) : $localize `:@@users.user-edit.created-on.label:Created on`),
-        'model.createdOn': (model) => (this.fieldsPropertyExpressions?.createdOn ? this.fieldsPropertyExpressions.createdOn(model) : model.createdOn),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['createdOn'] ? this.fieldsDisableExpressions['createdOn'](model) : true)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['createdOn'] ? this.fieldsRequiredExpressions['createdOn'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['createdOn'] ? this.fieldsLabelExpressions['createdOn'](model) : $localize `:@@users.user-edit.created-on.label:Created on`),
+        'model.createdOn': (model) => (this.fieldsPropertyExpressions?.['createdOn'] ? this.fieldsPropertyExpressions['createdOn'](model) : model.createdOn),
         },
         templateOptions: {
         label: $localize `:@@users.user-edit.created-on.label:Created on`,
         placeholder: $localize `:@@users.user-edit.created-on.label:Created on`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: { name: 'date' }
         },
 modelOptions: { updateOn: 'blur' },
@@ -233,19 +233,19 @@ modelOptions: { updateOn: 'blur' },
         type: this.resolveFieldType('datepicker', true),
         focus: false,
 className: `entry-updated-on-field entry-datepicker`,
-        hideExpression: this.fieldsHideExpressions?.updatedOn ?? false,
+        hideExpression: this.fieldsHideExpressions?.['updatedOn'] ?? false,
         expressionProperties: {
-        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.updatedOn ? this.fieldsDisableExpressions.updatedOn(model) : true)),
-        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.updatedOn ? this.fieldsRequiredExpressions.updatedOn(model) : false),
-        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.updatedOn ? this.fieldsLabelExpressions.updatedOn(model) : $localize `:@@users.user-edit.updated-on.label:Updated on`),
-        'model.updatedOn': (model) => (this.fieldsPropertyExpressions?.updatedOn ? this.fieldsPropertyExpressions.updatedOn(model) : model.updatedOn),
+        'templateOptions.disabled': (model) => (this.isReadonly || (this.fieldsDisableExpressions?.['updatedOn'] ? this.fieldsDisableExpressions['updatedOn'](model) : true)),
+        'templateOptions.required': (model) => (this.fieldsRequiredExpressions?.['updatedOn'] ? this.fieldsRequiredExpressions['updatedOn'](model) : false),
+        'templateOptions.label': (model) => (this.fieldsLabelExpressions?.['updatedOn'] ? this.fieldsLabelExpressions['updatedOn'](model) : $localize `:@@users.user-edit.updated-on.label:Updated on`),
+        'model.updatedOn': (model) => (this.fieldsPropertyExpressions?.['updatedOn'] ? this.fieldsPropertyExpressions['updatedOn'](model) : model.updatedOn),
         },
         templateOptions: {
         label: $localize `:@@users.user-edit.updated-on.label:Updated on`,
         placeholder: $localize `:@@users.user-edit.updated-on.label:Updated on`,
         description: ``,
             attributes: {  },
-            hidden: !true,
+        hidden: !true,
             typeFormatDef: { name: 'date' }
         },
 modelOptions: { updateOn: 'blur' },
@@ -255,10 +255,10 @@ modelOptions: { updateOn: 'blur' },
         ]
         },
     ];
-}
+    }
 
-  private applyOptionally<T>(value: T, apply: boolean): T | undefined {
+    private applyOptionally<T>(value: T, apply: boolean): T | undefined {
     return apply ? value : undefined;
-  }
+    }
 
-}
+    }
