@@ -8,7 +8,6 @@ import { FORM_FIELD_ERROR_KEY } from '@enigmatry/entry-components/validation';
 import { EntryFormConfig, EntryFormModule, ENTRY_FORM_CONFIG } from '@enigmatry/entry-form';
 import { FormlyModule } from '@ngx-formly/core';
 import { FormlyMaterialModule } from '@ngx-formly/material';
-import { EntryComponentsModule } from '@shared/entry-components.module';
 import { FormlyButtonComponent } from './formly-button/formly-button.component';
 import { FormlyExpansionPanelComponent } from './formly-expansion-panel/formly-expansion-panel.component';
 import { FormlyFieldsetComponent } from './formly-fieldset/formly-fieldset.component';
@@ -30,22 +29,19 @@ const defaultFormConfig: EntryFormConfig = {
 };
 
 @NgModule({
-  declarations: [
+  imports: [
     FormlyExpansionPanelComponent,
     FormlyFieldsetComponent,
     ReadonlyInputComponent,
     ReadonlyRadioComponent,
     ReadonlyBooleanComponent,
-    FormlyButtonComponent
-  ],
-  imports: [
+    FormlyButtonComponent,
     CommonModule,
     ReactiveFormsModule,
     MatExpansionModule,
     MatSelectModule,
     MatButtonModule,
     EntryFormModule,
-    EntryComponentsModule,
     FormlyMaterialModule,
     FormlyModule.forChild({
       types: [
