@@ -1,31 +1,28 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
+import { MatDateFnsModule } from '@angular/material-date-fns-adapter';
+import { provideMatDateFormats, provideMatDateLocale } from '@app/i18n/localization';
 import { FormlyExtensionsModule } from '../formly/formly-extensions.module';
 import { EntryComponentsModule } from './entry-components.module';
-import { FormWrapperComponent } from './form-wrapper/form-wrapper.component';
-import { GridCellsModule } from './grid-cells/grid-cells.module';
-import { MaterialModule } from './material.module';
-import { PipesModule } from './pipes/pipes.module';
 
 @NgModule({
   imports: [
     CommonModule,
-    MaterialModule,
     EntryComponentsModule
   ],
   exports: [
     FormsModule,
     ReactiveFormsModule,
-    MaterialModule,
-    GridCellsModule,
+    MatDateFnsModule,
     FormlyExtensionsModule,
-    EntryComponentsModule,
-    FormWrapperComponent,
-    PipesModule
+    EntryComponentsModule
   ],
-  declarations: [
-    FormWrapperComponent
+  providers: [
+    provideMatDateLocale(),
+    provideMatDateFormats(),
+    { provide: MAT_FORM_FIELD_DEFAULT_OPTIONS, useValue: { appearance: 'outline', floatLabel: 'auto' } }
   ]
 })
 export class SharedModule { }
